+++ map-matcher-library/src/main/java/nu/ndw/nls/routingmapmatcher/model/routing/RoutingRequest.java
@@ -1,6 +1,5 @@
 package nu.ndw.nls.routingmapmatcher.model.routing;
 
-import com.graphhopper.util.CustomModel;
 import java.util.List;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -13,9 +12,8 @@ import org.locationtech.jts.geom.Point;
 @EqualsAndHashCode
 @ToString
 public class RoutingRequest {
-    private final String routingProfile;
+
     private final List<Point> wayPoints;
-    private final CustomModel customModel;
     @Builder.Default
     private final boolean simplifyResponseGeometry = true;
     @Builder.Default